Ahora async_write incrementa el contador de datos pendientes de envío.

+++ network.h
@@ -36,7 +36,7 @@ struct epoll_data_client {
 };
 
 int aio_socket_escucha(int puerto);
-int async_write(struct epoll_data_client* data, void* buffer, ssize_t length);
+int async_write(struct epoll_data_client* data, void* buffer, int length);
 int async_write_delay(struct epoll_data_client* data);
 int async_read(struct epoll_data_client * data, void * buffer, int length);
 void init_epoll_data(int socketfd, struct epoll_data_client * data);
